backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
supabase==2.3.5
google-generativeai==0.8.0
boto3==1.29.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
PyJWT==2.8.1
